add jaeger and connect them with otlp
i think the problem was that i had no otlp pipeline from traces, so i
never saw anything in prometheus.

i added jaeger, which is good for viewing traces, and set up a pipeline
in otel-collector from incoming otlp to outgoing jaeger (still otlp).

i now see traces in jaeger, i see some metrics (i think) in prometheus,
and the logs are going to the console in otlp (and in the app).