add feature-flags controlled by appsettings
this will add the standard microsoft feature-management nuget package
and set up one feature-flag in appsettings that gives 50% of the requests
3 days of forecast and the rest 5 days of forecast.

this is intended as a demonstration of how to use feature-flags, and you
should remove this in your own project.

introduces a new class FeatureFlags that is used to access the feature-flags
from the rest of the application as constant strings.